# Implement IList<T> for EmptyPartition, RangeIterator and RepeatIterator.