--- a/src/System.Linq/src/Resources/Strings.resx
+++ b/src/System.Linq/src/Resources/Strings.resx
@@ -72,4 +72,7 @@
   <data name="NoMatch" xml:space="preserve">
     <value>Sequence contains no matching element</value>
   </data>
+  <data name="ArrayPlusOffTooSmall" xml:space="preserve">
+    <value>Destination array is not long enough to copy all the items in the collection. Check array index and length.</value>
+  </data>
 </root>