--- a/src/System.Linq/tests/EmptyPartitionTests.cs
+++ b/src/System.Linq/tests/EmptyPartitionTests.cs
@@ -106,5 +106,78 @@ namespace System.Linq.Tests
             en.Reset();
             en.Reset();
         }
+
+        [Fact]
+        public void ICollection_IsReadOnly()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as ICollection<int>;
+
+            Assert.Equal(true, emptyPartition.IsReadOnly);
+            Assert.Throws<NotSupportedException>(() => emptyPartition.Add(0));
+            Assert.Throws<NotSupportedException>(() => emptyPartition.Remove(0));
+            Assert.Throws<NotSupportedException>(() => emptyPartition.Clear());
+        }
+
+        [Fact]
+        public void ICollection_Count()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as ICollection<int>;
+
+            Assert.Equal(0, emptyPartition.Count);
+        }
+
+        [Fact]
+        public void ICollection_CopyTo_ProduceCorrectSequence()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as ICollection<int>;
+
+            var arrayIndex = 5;
+            var array = Enumerable.Range(0, 10).ToArray();
+            emptyPartition.CopyTo(array, arrayIndex);
+            
+            for (var index = 0; index < 10; index++)
+            {
+                Assert.Equal(index, array[index]);
+            }
+        }
+
+        [Fact]
+        public void ICollection_Contains()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as ICollection<int>;
+
+            Assert.Equal(false, emptyPartition.Contains(int.MinValue));
+            Assert.Equal(false, emptyPartition.Contains(0));
+            Assert.Equal(false, emptyPartition.Contains(int.MaxValue));
+        }
+
+        [Fact]
+        public void IList_IsReadOnly()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as IList<int>;
+
+            Assert.Throws<NotSupportedException>(() => emptyPartition.Insert(0, 0));
+            Assert.Throws<NotSupportedException>(() => emptyPartition.RemoveAt(0));
+        }
+
+        [Fact]
+        public void IList_Indexer_ThrowExceptionOnOutOfRangeIndex()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as IList<int>;
+
+            AssertExtensions.Throws<ArgumentOutOfRangeException>("index", () => emptyPartition[int.MinValue]);
+            AssertExtensions.Throws<ArgumentOutOfRangeException>("index", () => emptyPartition[0]);
+            AssertExtensions.Throws<ArgumentOutOfRangeException>("index", () => emptyPartition[int.MaxValue]);
+        }
+
+        [Fact]
+        public void IList_IndexOf()
+        {
+            var emptyPartition = GetEmptyPartition<int>() as IList<int>;
+
+            Assert.Equal(-1, emptyPartition.IndexOf(int.MinValue));
+            Assert.Equal(-1, emptyPartition.IndexOf(0));
+            Assert.Equal(-1, emptyPartition.IndexOf(int.MaxValue));
+        }    
     }
 }