## `Empty`

`Enumerable.Empty<T>` has two implementations. The `SizeOpt` version returns `Array.Empty<T>`, while the `SpeedOpt` version returns `EmptyPartition<T>.Instance`. 

Besides the different optimizations, there's another difference: `Array.Empty<T>` returns an implementation of `IList<T>`, while `EmptyPartition<T>.Instance` returns an implementation of `IEnumerable<T>`.

This difference may cause some confusion as it behaves differently in different platforms, not only between different versions, as already reported in #32645.

@natemcmaster This PR adds the implementation of `IList<T>` as discussed on #32645. It also adds `IReadOnlyList<T>`.

## `Range` and `Repeat`

These LINQ operations return implementations of `Iterator<T>` which implements `IEnumerable<T>`. But, for both these operations, `Count` is well known and the items can be inferred from the `index` value. This means that these iterators can implement `IList<T>`.

The advantage is that `ICollection<T>` and `IList<T>` allow much better performance than `IEnumerable<T>` in many situations. Many optimizations based on these interfaces can be found in this same repository and in many third-parties.

Some of these cases may be overloaded by even better optimizations (`SpeedOpt` and `SizeOpt`) but these are not used in all platforms. Still, these changes, take advantage of existing code. No new cases are created.

This PR adds implementation of `IList<T>` and `IReadOnlyList<T>` to the iterators `RangeIterator` and `RepeatIterator<T>`.

Both `Range` and `Repeat` return `Empty<T>()` when `count` is zero. That's one other reason why it's important `EmptyIterator<T>` also implements `IList<T>`, reducing confusion.

## `ToList()`

One of the cases that is overloaded in `SpeedOpt`, is the `ToList()`method for both `RangeIterator` and `RepeatIterator<T>`. They simply create a `List<T>` and add the elements to it. Here's the [implementation for Range](https://github.com/dotnet/corefx/blob/81c8c68d81daad033251d643e230061f38a9ae3a/src/System.Linq/src/System/Linq/Range.SpeedOpt.cs#L31):

```csharp
public List<int> ToList()
{
   List<int> list = new List<int>(_end - _start);
   for (int cur = _start; cur != _end; cur++)
   {
       list.Add(cur);
   }

   return list;
}
```

Although the length is passed as the `capacity` for the new `List<T>`, the `Add()` method still performs version increment, buffer length validation and size increment. This is performed for every element, which is unnecessary and penalizing.

```csharp
public void Add(T item)
{
    _version++;
    T[] array = _items;
    int size = _size;
    if ((uint)size < (uint)array.Length)
    {
        _size = size + 1;
        array[size] = item;
    }
    else
    {
        AddWithResize(item);
    }
}
```

As an alternative, the `List<T>(IEnumerable<T>)` constructor can be used, which has the following implementation:

```csharp
public List(IEnumerable<T> collection)
{
    if (collection == null)
        ThrowHelper.ThrowArgumentNullException(ExceptionArgument.collection);

    if (collection is ICollection<T> c)
    {
        int count = c.Count;
        if (count == 0)
        {
            _items = s_emptyArray;
        }
        else
        {
            _items = new T[count];
            c.CopyTo(_items, 0);
            _size = count;
        }
    }
    else
    {
        _size = 0;
        _items = s_emptyArray;
        using (IEnumerator<T> en = collection.GetEnumerator())
        {
            while (en.MoveNext())
            {
                Add(en.Current);
            }
        }
    }
}
```

When `collection` is a `ICollection<T>`, which is now the case, it simply allocates an array, calls `ICollection<T>.CopyTo()` and sets the size. 

This PR replaces the `for` loop with `List.Add()` calls by a single call to the `List<T>(IEnumerable<T>)` constructor.

## Benchmarks

These are the benchmarks for `Range` with multiple count values (0, 1, 10, 100, 1.000, 10.000). The first line values are for `master` branch and the second for this PR branch.

``` ini

BenchmarkDotNet=v0.11.3.1003-nightly, OS=macOS Mojave 10.14.4 (18E226) [Darwin 18.5.0]
Intel Core i5-7360U CPU 2.30GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.0.100-preview4-011223
  [Host]     : .NET Core 3.0.0-preview4-27615-11 (CoreCLR 4.6.27615.73, CoreFX 4.700.19.21213), 64bit RyuJIT
  Job-CFRYCL : .NET Core e70c7a22-15eb-4e75-85bf-9200f5f40b1a (CoreCLR 4.6.27701.72, CoreFX 4.700.19.25201), 64bit RyuJIT
  Job-XREZYG : .NET Core 0ac558bb-ff95-4358-9e42-24ac416bfb32 (CoreCLR 4.6.27624.71, CoreFX 4.700.19.23001), 64bit RyuJIT

Runtime=Core  IterationTime=250.0000 ms  MaxIterationCount=20  
MinIterationCount=15  WarmupCount=1  

```
| Method |                                                                                                     Toolchain | Count |         Mean |      Error |     StdDev |       Median |          Min |          Max | Ratio | MannWhitney(10%) | RatioSD | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |-------------------------------------------------------------------------------------------------------------- |------ |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|------:|----------------- |--------:|------------:|------------:|------------:|--------------------:|
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |     **0** |     **20.25 ns** |  **1.2950 ns** |  **1.4913 ns** |     **20.13 ns** |     **18.46 ns** |     **22.86 ns** |  **1.04** |             **Same** |    **0.08** |      **0.1958** |           **-** |           **-** |                **32 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |     0 |     19.47 ns |  0.5175 ns |  0.5959 ns |     19.28 ns |     18.77 ns |     20.66 ns |  1.00 |             Base |    0.00 |      0.1958 |           - |           - |                32 B |
|        |                                                                                                               |       |              |            |            |              |              |              |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |     **1** |     **52.89 ns** |  **0.6164 ns** |  **0.5464 ns** |     **52.76 ns** |     **52.07 ns** |     **53.81 ns** |  **1.45** |           **Slower** |    **0.02** |      **0.3917** |           **-** |           **-** |                **64 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |     1 |     36.52 ns |  0.2953 ns |  0.2762 ns |     36.45 ns |     36.23 ns |     37.18 ns |  1.00 |             Base |    0.00 |      0.3916 |           - |           - |                64 B |
|        |                                                                                                               |       |              |            |            |              |              |              |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |    **10** |     **63.85 ns** |  **0.5476 ns** |  **0.4855 ns** |     **63.83 ns** |     **63.10 ns** |     **64.83 ns** |  **1.07** |             **Same** |    **0.02** |      **0.5873** |           **-** |           **-** |                **96 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |    10 |     59.60 ns |  0.9767 ns |  0.8658 ns |     59.39 ns |     58.63 ns |     61.37 ns |  1.00 |             Base |    0.00 |      0.5874 |           - |           - |                96 B |
|        |                                                                                                               |       |              |            |            |              |              |              |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |   **100** |    **200.93 ns** |  **2.2097 ns** |  **1.9589 ns** |    **200.56 ns** |    **198.80 ns** |    **205.87 ns** |  **0.69** |           **Faster** |    **0.04** |      **2.7855** |           **-** |           **-** |               **456 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |   100 |    289.58 ns | 19.1209 ns | 18.7793 ns |    281.08 ns |    277.56 ns |    331.04 ns |  1.00 |             Base |    0.00 |      2.7853 |           - |           - |               456 B |
|        |                                                                                                               |       |              |            |            |              |              |              |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** |  **1000** |  **1,450.43 ns** | **14.6100 ns** | **13.6662 ns** |  **1,449.04 ns** |  **1,423.05 ns** |  **1,477.32 ns** |  **0.60** |           **Faster** |    **0.01** |     **24.3859** |           **-** |           **-** |              **4056 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun |  1000 |  2,426.86 ns | 22.6250 ns | 18.8929 ns |  2,420.40 ns |  2,397.13 ns |  2,461.57 ns |  1.00 |             Base |    0.00 |     24.3898 |           - |           - |              4056 B |
|        |                                                                                                               |       |              |            |            |              |              |              |       |                  |         |             |             |             |                     |
| **ToList** |          **/corefx/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun** | **10000** |  **8,182.43 ns** | **97.2001 ns** | **90.9211 ns** |  **8,176.24 ns** |  **8,079.06 ns** |  **8,427.27 ns** |  **0.46** |           **Faster** |    **0.01** |     **57.1382** |      **7.1382** |           **-** |             **40056 B** |
| ToList | /corefx_upstream/artifacts/bin/testhost/netcoreapp-OSX-Release-x64/shared/Microsoft.NETCore.App/9.9.9/corerun | 10000 | 17,766.04 ns | 98.3112 ns | 91.9603 ns | 17,780.30 ns | 17,573.70 ns | 17,901.24 ns |  1.00 |             Base |    0.00 |     57.1206 |      7.0781 |           - |             40056 B |

Please note that when `count` is zero, `Range` return `Empty<int>()`. The new version has the same performance.

For the other cases where `RangeIterator` is returned, there is an initialization overhead due to the cast and argument validation. It's noticeable for very small values of `count`. It's 45% times slower when `count` is 1. 

This overhead dilutes quickly when `count` increases. It takes more or less the same time when `count` is 10. It's 30% faster when `count` is 100, 40% faster when `count` is 1.000 and  54% faster when `count` is 10.000.

In absolute values, this means that, it's only 16ns slower when `count` is 1 but almost 10.000ns faster when `count` is 10.000.

There's no memory overhead added.