@stephentoub 2019-05-03
> as already reported in #32645

Why would implementing `IList<T>` be sufficient to avoid this "confusion"?  Can you share some examples where developers are actually confused by this?  If we truly believe that's a problem, why do we believe folks expect it to implement `IList<T>` but not actually be a `T[]`?

We make no guarantees about the concrete type that's returned from these APIs, nor what interfaces they may or may not implement.  Anyone who's casting should be doing so with a type check and a fallback.

But if we were going to make a change to `Empty<T>` because of concerns about the concrete type returned and its interface implementations, then wouldn't we just go back to using `Array.Empty<T>`? I don't think it makes sense to start adding back such interfaces piecemeal, and even then not actually capture the full expectation and avoid that confusion.

> The advantage is that ICollection<T> and IList<T> allow much better performance than IEnumerable<T> in many situations. Many optimizations based on these interfaces can be found in this same repository and in many third-parties.

I just spent half an hour looking through a bunch of code for use of Enumerable.Range(...) and Enumerable.Repeat(...).  I found very few cases where they were directly followed by ToList and where performance mattered.  At least in the code I looked at, by far the most common thing to follow Range was a Select, so I opened https://github.com/dotnet/corefx/pull/37410.  After that was ToArray, which is already handled by the RangeIterator implementation.  Enumerable.Repeat itself is very rare in anything but test code I looked at.

> but these are not used in all platforms

They're not used in all platforms to avoid a bunch of additional generic instantiations and interface implementations.  Adding `IList<T>` implementations has the same problem.