@aalmada 2019-05-03
> LINQ is an enticing area to add optimizations, with practically an infinite number of them possible. But it's infeasible for the implementation to provide optimizations for every possible combination of operator and input source, in terms of implementation and subsequent code maintainence, in terms of binary size, and in terms of the tradeoffs that result, and so we have basic implementations that work with everything and then only special-case those patterns we've found to be common and that are expected to make a meaningful difference on consuming code, and in particular in consuming code where such improvements actually matter. I'm not convinced this meets that bar. If you have real examples where it would, please do share; I'm happy to be convinced otherwise.

@stephentoub I'm well aware that LINQ is a beast. I've been trying to contribute for while but it's been hard to find a spot where it doesn't break everything. Where, it can radically improves big data without affecting tiny collections. 

I'm also well aware that Range().ToArray() or Range().Select().ToArray() are way more common than Range().ToList(). This was a first attempt of a simple contribution, for the sake of code completeness and to evaluate acceptance criteria.

I would think that #32645 is an example of someone getting confused. Also, I hope my comment there clarifies my opinion.

I've been working on [my own "green field" version of LINQ](https://github.com/NetFabric/NetFabric.Hyperlinq) where I test my own ideas. I'm not worried about binary size but with performance on big collections.

Looks like I'm going to stick to my work only but you're welcome to comment there.