@stephentoub 2019-05-03
> I've been trying to contribute for while but it's been hard to find a spot where it doesn't break everything. Where, it can radically improves big data without affecting tiny collections.

The contributions really are appreciated.  As you suggest, though, it can be difficult to find places to make meaningful improvements.  For several years we had a few contributors actively optimizing various parts of LINQ, and so lots of the low-hanging fruit has already been picked.  Optimizations at this point generally involve tradeoffs that need to be scrutizined or alternatively necessitate data to demonstrate why the cost (e.g. extra code) is worth it in terms of the payoff.

> I would think that #32645 is an example of someone getting confused. Also, I hope my comment there clarifies my opinion.

The question there was, "is it reasonable for `(IList<T>)Empty<T>()` to keep working?"  The same question could be asked with "is it reasonable for `(T[])Empty<T>()` to keep working?"  I think the answer needs to be the same, as if folks are expecting the type to be the same as it was previously, then I don't know why that would apply to a subset of what it implemented.

> Looks like I'm going to stick to my work

Ok.  Best of luck with your project.  If you find things that you think would be valuable contributions back to System.Linq, we'd welcome the discussion.

Thanks!