@GSPP 2019-05-05
When `Empty<T>()` was made to return an array it was entirely predictable that people would (mistakenly) take a dependency on that. It is generally my view that the first implementation in the framework should intentionally try to break such dependencies as much as possible. `Empty<T>()` should have taken care to *not* return anything deriving from `IList`.

Just as an example, changing `String.GetHashCode()` was hard because it was deterministic in 1.0. It would have been better to randomize it in 1.0 using a very simple scheme (e.g. `return actualHash ^ staticHash;`) to break people from taking a dependency.

If `Empty<T>()` actually returns an `IList` this becomes effectively contractual. Therefore, the method is pointless. People can directly use `Array.Empty<T>()`. Or, the return type could directly be made to promise `IList`.